Removed server dependencies & refactoring of client
*General changes*
Removed all server dependencies from the client. This breaks some of the semantics of the client code as it won't serve the same purpose as it did before. For example: logout() is now only used to close the application.

*Timer refactor*
Refactored timer to use a timer based on seconds instead of minutes and seconds. This results in a lot less complexity of the code while achieving the same results.

*Postponed activity & ActivityManager*
Implemented ability to postpone activity when pressing the postpone button in the notification. To achieve this, a queue is added to ActivityManager. This queue consists of the postponed activities. As such getActivity() will remove an item from this queue if it is not empty - else it will get a random activity as usual.

*GUI changes*
Removed the change timer button as it introduced a couple of bugs when using the application. There's now only a start timer button which will kill all existing timers and start a new one with the specified timer interval. This might be changed in the future if the team considers this design to be not be good UX.