Modified Activities to save ImageIcon & ImagePath
As Java Swing ImageIcons can't be serialised we have to save the path to the image to later reinitialise them later.

Changed activity IDs to UUIDs as the previous system with ints IDs based on the activity list size would have lead to inconsistencies if a user were to delete an activity that wasn't the last one in the list and then add a new one. The new activity would've then gotten an ID of an existing activity.

Lastly, customised hashcode/equals methods to exclude the ImageIcon as it is dynamic. If this wasn't in place, tests verifying the saving/loading of activity objects would not get a match as the two ImageIcons would point to different ImageIcon objects. That is: the two activity objects would have been the same with the only difference being the memory location of the ImageIcon object.