Finished the implementation of saving of activities to disk
*Moved to JSON-based serialisation*
The old method of saving/reading from disk was fragile and undocumented. As such I made the decision to use a more standardised approach by using a JSON serialisation library, in this case Gson. It works by saving objects to JSON format and then when reading it builds Java objects from the JSON. This approach has numerus benefits. The first being that the data is human readable as it contains a key, e.g. "activityName" and a value, e.g. "Situps". The second reason is that this approach supports easily adding new attributes in the future to the activity class without breaking the reading/writing methods.

*Error handling*
Improved error handling when reading/writing files. Also added suitable error messages that also points to the current method to aid in locating potential errors.

*Miscellaneous changes*
Added the helper methods fileExists() and createFile() to make the code more readable. Also added wrapper methods for save/load activities that uses the already set up activitiesFilePath in ActivityManager.