Add method addCustomActivity & refactor notification formatting
*Custom activities*
Added the ability to add custom activities, and in doing that - the ability to add descriptions to custom activities (hard to implement adding descriptions to custom activities without being able to add custom activities). When the user presses the "add new activity" button, the user is prompted to enter an activity name, instructions, and information about the activity. Optionally, an image can be added as well.

*HTML formatting*
As it was difficult to integrate the previously used bespoke formatting "language" into the "add custom activity"-panel, it has now been replaced. The replacement is simply displaying the activity instructions in HTML with some CSS styling to handle line break formatting automatically. This is easier to use and understand.

*Misc*
Changed return types of getActivity methods to Optionals to be able to handle no activity being returned.

Added some javadoc comments for the more important methods in AppPanel.